+++ fxmanifest.lua
@@ -5,8 +5,8 @@ lua54 'yes'
 
 name 'fks-hud'
 author 'Fonkas'
-description 'HUD + metabolism + consumables for RSG Framework'
-version '1.0.0'
+description 'HUD + metabolism + consumables for RedM (RSG and VORP)'
+version '1.1.0'
 license 'GPL-3.0-or-later' -- see LICENSE
 -- fks-hud  Copyright (C) 2026 Fonkas
 
@@ -16,9 +16,11 @@ shared_scripts {
     'config_items.lua',
     'config_animations.lua',
     'locales/*.lua',
+    'bridge/shared.lua',
 }
 
 client_scripts {
+    'bridge/client.lua',
     'client/utils.lua',
     'client/hud.lua',
     'client/metabolism.lua',
@@ -28,6 +30,7 @@ client_scripts {
 }
 
 server_scripts {
+    'bridge/server.lua',
     'server/main.lua',
     'server/version.lua',
 }
@@ -42,8 +45,8 @@ files {
     'html/img/*',
 }
 
+-- plus rsg-core or vorp_core (+ vorp_inventory), detected automatically (Config.Framework)
 dependencies {
-    'rsg-core',
     'ox_lib',
     'oxmysql',
 }